% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{applyActiveThreshold}
\alias{applyActiveThreshold}
\title{Remove predictors following an activation threshold}
\usage{
applyActiveThreshold(data, threshold = 0.1, perc = 0.8)
}
\arguments{
\item{data}{numerical matrix with predictors as rows and samples as columns}

\item{threshold}{numeric: minimum TPM value to consider a predictor 'active'}

\item{perc}{numeric: minimum percentage of activated predictors in all
samples.}
}
\value{
reduced numerical matrix
}
\description{
Identifies and eliminates predictors that do not fulfill the activation
threshold requirement. By default, a predictor should have a value greater
that 0.1 in at least the 80\% of the samples.
}
