% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpreproc.R
\name{applyCorrSkew}
\alias{applyCorrSkew}
\title{Remove correlated predictors}
\usage{
applyCorrSkew(data, cutoff = 0.9)
}
\arguments{
\item{data}{numerical matrix with predictors as rows and samples as columns}

\item{cutoff}{numeric: value for the pair-wise absolute correlation cutoff
(\link[caret]{findCorrelation}).}
}
\value{
reduced numerical matrix
}
\description{
Identifies and eliminates one of two predictors that have high correlation
(>= 90\%).
}
