% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{applyLog2}
\alias{applyLog2}
\title{Log2 + 1 to all matrix}
\usage{
applyLog2(data)
}
\arguments{
\item{data}{non-negative numerical matrix}
}
\value{
numerical matrix
}
\description{
Applies a factor log2 + 1 to all values in the data matrix.
}
