% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{applyNearZeroVar}
\alias{applyNearZeroVar}
\title{Remove near zero variance predictors}
\usage{
applyNearZeroVar(data, freqCut = 20, uniqueCut = 5, sigDigits = 1)
}
\arguments{
\item{data}{numerical matrix with predictors as rows and samples as columns}

\item{freqCut}{numerical, the cutoff for the ratio of the most common value to the second most common value}

\item{uniqueCut}{numerical, the cutoff for the percentage of distinct values out of the number of total samples}

\item{sigDigits}{numerical, number of significant digits to round}
}
\value{
reduced numerical matrix
}
\description{
Identifies and eliminates predictors that have very few unique values and the
ratio of the frecuency of the most common value to the frequency of the
second most common value is large. More info at \link[caret]{nearZeroVar}.
}
