% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{applyZScore}
\alias{applyZScore}
\title{Scaling, centering and normalizing a matrix}
\usage{
applyZScore(data, by.columns = FALSE)
}
\arguments{
\item{data}{non-negative numerical matrix with column names}

\item{by.columns}{boolean. Whether to normalize by columns of rows, by default FALSE}
}
\value{
numerical matrix
}
\description{
Scales a non-negative numerical matrix by centering and normalizing its data.
}
