% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSelection.R
\name{bestSeed}
\alias{bestSeed}
\title{Calculates the best seed for GLMNET algorithm}
\usage{
bestSeed(data, covariate, useRMSE = FALSE)
}
\arguments{
\item{data}{expression matrix}

\item{covariate}{numeric vector}

\item{useRMSE}{boolean parameter}
}
\value{
a number representing the best seed
}
\description{
Calculates the best seed for GLMNET algorithm
}
