% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{boot.RMSE}
\alias{boot.RMSE}
\title{Bar Plot of the RMSE values for each tissue}
\usage{
boot.RMSE(tissues, dataset_folder, models_folder, B = 20)
}
\arguments{
\item{tissues}{list of strings containing the tissue names}

\item{dataset_folder}{path to the dataset folder}

\item{models_folder}{path to the models folder}

\item{B}{number of bootstrap resamples (default 20)}
}
\value{
plot
}
\description{
Bar Plot of the RMSE values for each tissue
}
