% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryf.R
\name{calculateCorrelation}
\alias{calculateCorrelation}
\title{Calculate correlation}
\usage{
calculateCorrelation(x, y, tam)
}
\arguments{
\item{x}{coexpression matrix}

\item{y}{factor}

\item{tam}{number}
}
\value{
data.frame
}
\description{
Given a coexpression matrix x and a gene (which is represented as a numerical vector) y from that matrix,
this function returns a subset of x with the genes most correlated with y.
The size of this subset is tam
}
