% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{calculateSimulationConfussionMatrix}
\alias{calculateSimulationConfussionMatrix}
\title{Generates the confussion matrix for a simulation}
\usage{
calculateSimulationConfussionMatrix(set.calculated, sim.obj, covariate = "Q1")
}
\arguments{
\item{set.calculated}{vector of predictors selected by the model.}

\item{sim.obj}{simulation object obtained from CovCoExpNets.}

\item{covariate}{which covariate to study. Either "Q1" or "Q2". Defaults
to Q2.}
}
\value{
caret confussion matrix.
}
\description{
Given a set of predictors selected by the model (set.calculated)
and the simulation object to extract the total number of predictors,
this function generates the confussion matrix when discriminating
relevant vs non-relevant predictors.
}
