% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSelection.R
\name{coexpressionNetwork}
\alias{coexpressionNetwork}
\title{Co-expression network}
\usage{
coexpressionNetwork(data, genes)
}
\arguments{
\item{data}{expression matrix}

\item{genes}{dataframe with the genes selected by the glmnet algorithm and their coefficients}
}
\value{
A list where for each gene selected by glmnet appears the 50 genes most correlated with it and that correlation
}
\description{
Calculate a coexpression network using the genes selected by the glmnet algorithm
as the best predictors of a covariate.
}
