% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSelection.R
\name{coexpressionNetworkFixed}
\alias{coexpressionNetworkFixed}
\title{Co-expression network with fixed cluster sizes}
\usage{
coexpressionNetworkFixed(data, selected.genes, tam)
}
\arguments{
\item{data}{expression matrix}

\item{selected.genes}{list of dataframes with the genes selected as important
by GLMNET algorithm}

\item{tam}{number: indicates the number of genes in each cluster}
}
\value{
A dataframe where for each gene selected by glmnet appears the tam
genes most correlated with it and that correlation. If only one object was
supplied, the output will be one object and not a list.
}
\description{
Calculate a coexpression network using the genes selected by the glmnet
algorithm as the best predictors of a covariate. The inputs can be either a
list object generated from previous SuCoNets functions or individual
condition variables.
}
