% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSelection.R
\name{coexpressionNetworkVariable}
\alias{coexpressionNetworkVariable}
\title{Co-expression network with variable cluster sizes}
\usage{
coexpressionNetworkVariable(data, covariate, selected.genes)
}
\arguments{
\item{data}{list of expression matrices}

\item{covariate}{list of numeric vectors}

\item{selected.genes}{list of dataframes with the genes selected as important
by GLMNET algorithm}
}
\value{
A dataframe where for each gene selected by glmnet appears the tam
genes most correlated with it and a vector with cluster sizes. If only one
object was supplied, the output will be one object and not a list.
}
\description{
Calculate a coexpression network using the genes selected by the glmnet
algorithm as the best predictors of a covariate. The inputs can be either a
list object generated from previous SuCoNets functions or individual
condition variables.
}
