% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{comparisonActualPredictedCovariate}
\alias{comparisonActualPredictedCovariate}
\title{Comparison between real values of the covariate and predicted values}
\usage{
comparisonActualPredictedCovariate(data, covariate, m, d, cvfit, seed)
}
\arguments{
\item{data}{expression matrix}

\item{covariate}{numeric vector}

\item{m}{mean of the original covariate}

\item{d}{standard deviation of the original covariate}

\item{cvfit}{object generated with GLMNET algorithm}

\item{seed}{number}
}
\value{
plot
}
\description{
Comparison between real values of the covariate and predicted values
}
