% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{convertToNamedList}
\alias{convertToNamedList}
\title{Transforms a variable into a named list object}
\usage{
convertToNamedList(data.return, names.reference, iter, return.list)
}
\arguments{
\item{data.return}{data generated inside the loop.}

\item{names.reference}{names of the data from which the data.return variable was generated.}

\item{iter}{iteration of the loop.}

\item{return.list}{whether the input of the main function was a list or not.}
}
\value{
named list containing the generated variable.
}
\description{
Usually employed to generate the return list by the \link[foreach]{foreach}
loops.
}
