% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{correctGTExNames}
\alias{correctGTExNames}
\title{Corrects GTEx naming convention}
\usage{
correctGTExNames(data)
}
\arguments{
\item{data}{numerical matrix with predictors as rows and samples as columns}

\item{df.bool_data}{boolean matrix with samples as rows and conditions as
columns. The data matrix will be splitted according to this boolean matrix}

\item{names}{names assigned to each different numerical matrix in the
returned list. If not provided, will be selected from the 'df.bool_data'
matrix}
}
\value{
named list of numerical matrices
}
\description{
Removes the version term of the ENSEMBL naming convention of GTEx data.
More information \href{https://www.ensembl.org/Help/Faq?id=488}{here}
}
