% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpreproc.R
\name{dataPreprocess}
\alias{dataPreprocess}
\title{Preprocess the input data matrices}
\usage{
dataPreprocess(
  data,
  doSCN = TRUE,
  doNZV = TRUE,
  doAT = TRUE,
  doCorr = TRUE,
  threshold = 0.1,
  perc = 0.8,
  cutoff = 0.9,
  correctGTExPortal = TRUE
)
}
\arguments{
\item{data}{numerical matrix (or list of numerical matrices) with predictors
as rows and samples as columns}

\item{doSCN}{boolean to apply \link[SuCoNets]{scn}}

\item{doNZV}{boolean to apply \link[SuCoNets]{applyNearZeroVar}}

\item{doAT}{boolean to apply \link[SuCoNets]{applyActiveThreshold}}

\item{doCorr}{boolean to apply \link[SuCoNets]{applyCorrSkew}}

\item{threshold}{numeric: minimum TPM value (after applying
\link[SuCoNets]{scn}) to consider a predictor 'active'. Defaults to 0.1}

\item{perc}{numeric: minimum percentage of activated predictors in all
samples. Defaults to 0.8}

\item{cutoff}{numeric: value for the pair-wise absolute correlation cutoff
(\link[caret]{findCorrelation}). Defaults to 0.9}

\item{correctGTExPortal:}{boolean to apply a correction to the GTExPortal
gene naming convention. Defaults to True}
}
\value{
named list of reduced numerical matrices. If only one data matrix is
provided, it will return the reduced matrix, not a list.
}
\description{
Applies all the preprocessing required for the SuCoNets pipeline to work.
Each individual process can be deactivadted (but not recommended). It also
includes the possibility of applying a Gene name correction for GTExPortal
data.
}
