% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpreproc.R
\name{dataSplit}
\alias{dataSplit}
\title{Splits data matrix}
\usage{
dataSplit(data, df.bool_data, names)
}
\arguments{
\item{data}{numerical matrix with predictors as rows and samples as columns}

\item{df.bool_data}{boolean matrix with samples as rows and conditions as
columns. The data matrix will be splitted according to this boolean matrix}

\item{names}{names assigned to each different numerical matrix in the
returned list. If not provided, will be selected from the 'df.bool_data'
matrix}
}
\value{
named list of numerical matrices
}
\description{
Creates a lists of numeric matrices. Each matrix corresponds to a column in
the 'df.bool_data' matrix. The names of the conditions can be provoided or
extracted automatically from the 'df.bool_data' column names.
}
