% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSelection.R
\name{detectGenes}
\alias{detectGenes}
\title{Detect main genes}
\usage{
detectGenes(data, covariate)
}
\arguments{
\item{data}{expression matrix}

\item{covariate}{numeric vector}
}
\value{
Dataframe containing the variables selected by the glmnet algorithm and the coefficients associated with these variables
}
\description{
Given an expression matrix and a covariate, this functioin calculates the variables
that best predict that covariate using glmnet algorithm
}
