% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{distributionIndividualsCovariate}
\alias{distributionIndividualsCovariate}
\title{Distribution of individuals according to the covariate}
\usage{
distributionIndividualsCovariate(data, covariate, selected.genes, m, d)
}
\arguments{
\item{data}{list of expression matrices}

\item{covariate}{list of numeric vectors}

\item{selected.genes}{list of dataframes with the genes selected as important
by GLMNET algorithm}

\item{m}{list of meas of the original covariates}

\item{d}{list of standard deviations of the original covariates}
}
\value{
plot
}
\description{
The inputs can be either a list object generated from previous SuCoNets
functions or individual condition variables
}
\examples{
distributionIndividualsCovariate(data, covariate, selected.genes, m, d)
distributionIndividualsCovariate(data[[1]], covariate[[1]], selected.genes[[1]], m[[1]], d[[1]])
}
