% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{evaluateModel}
\alias{evaluateModel}
\title{Evaluates the CovCoExpNets models}
\usage{
evaluateModel(
  cvfit,
  data.test,
  covariate.test,
  genes.subset,
  data.train = NA,
  covariate.train = NA,
  m = NA,
  d = NA
)
}
\arguments{
\item{cvfit}{list of models generated with \link[CovCoExpNets]{glmnetGenesSubset}}

\item{data.test}{list of numeric expression matrices for testing}

\item{covariate.test}{list of numeric vectors for training}

\item{genes.subset}{list of selected predictors for each condition}

\item{data.train}{list of numeric expression matrices for training}

\item{m}{list of means for the covariate}

\item{d}{list of standard deviations for the covariate}
}
\value{
The calculated adjuted r^2
}
\description{
Given the models, testing datasets and the subset of genes, this
function generates the main metrics RMSE, returned predictors, r2
and adjusted r2.
}
\details{
If provided with the train dataset, it also calculates the metrics
for it.

The optional parameters "m" and "d" represents the mean and standard
deviation of the input covariate as returned by \link[CovCoExpNets]{normalize}
}
