% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{evaluateModelGLMNET}
\alias{evaluateModelGLMNET}
\title{Generate and evaluates GLMNET models}
\usage{
evaluateModelGLMNET(
  data.train,
  covariate.train,
  data.test,
  covariate.test,
  glmnet.family = "gaussian",
  m = NA,
  d = NA
)
}
\arguments{
\item{data.train}{list of numeric expression matrices for training}

\item{data.test}{list of numeric expression matrices for testing}

\item{covariate.test}{list of numeric vectors for testing}

\item{glmnet.family}{see \link[glmnet]{cv.glmnet} family parameter.}

\item{m}{list of means for the covariate}

\item{d}{list of standard deviations for the covariate}
}
\value{
data.frame with the most relevant metrics for each condition given
}
\description{
From a given train and test datasets, it generates the GLMNET models
and evaluates the most relevant metrics: RMSE, returned predictors, r2
and adjusted r2.
}
\details{
The optional parameters "m" and "d" represents the mean and standard
deviation of the input covariate as returned by \link[CovCoExpNets]{normalize}
}
