% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{extractGenesIter}
\alias{extractGenesIter}
\title{Extracts the relevant genes for a particular iteration}
\usage{
extractGenesIter(gene.freq, iteration)
}
\arguments{
\item{gene.freq}{data.frame containing the GLMNET repetitions results.
Should be the output of \link[CovCoExpNets]{geneFrequency}}

\item{iter}{numeric, particular to iteration to extract relevant
genes from.}
}
\value{
the genes selected as relevant for a particular iteration.
}
\description{
Extracts the relevant genes for a particular iteration
}
