% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubGenesDetection.R
\name{geneFrequency}
\alias{geneFrequency}
\title{GLMNET repetitions}
\usage{
geneFrequency(
  data,
  covariate,
  t = 10,
  k.folds = 10,
  train.split = 1,
  sample.prob = c(),
  seed = sample(1:999999, 1),
  glmnet.family = "gaussian"
)
}
\arguments{
\item{data}{list of numeric expression matrices}

\item{covariate}{list of numeric vectors}

\item{t}{number of times to repeat the train/test split. Defaults to 10}

\item{k.folds}{number of kfolds to execute in the GLMNET algorithm
(\link[glmnet]{cv.glmnet}). Defaults to 10}

\item{train.split}{numeric percentage of samples to take as train}

\item{sample.prob}{list of numeric vectors as weights when applying
\link[base]{sample}}

\item{seed}{fixed seed to generate split the dataset in train/test}

\item{glmnet.family}{see \link[glmnet]{cv.glmnet} family parameter.}
}
\value{
list of vectors containing the genes selected for each condition. If
only one data matrix is provided, it will return the reduced matrix, not a
list.
}
\description{
Executes GLMNET a set number of times (t) and extracts the genes selected as
relevant and its coefficient. The inputs can be either a list object
generated from previous CovCoExpNets functions or individual condition variables.
}
\details{
To measure the RMSE per iteration, please refer to the \link[CovCoExpNets]{geneFrequencyRMSE}
function.
}
