% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{generateFormula}
\alias{generateFormula}
\title{Generate simulation formula}
\usage{
generateFormula(P, start = 1)
}
\arguments{
\item{P}{number of predictors to simulate.}

\item{start}{numberic, where to begin the formula. By default, 1}
}
\value{
linear formula that the simulation will follow
}
\description{
Generate the formula that will follow the simulations. It will produce a linear
relationship between the response variable y and the predictors x:
y ~ B1\emph{x1 + B2}x2
}
