% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSelection.R
\name{geneSelection}
\alias{geneSelection}
\title{Reduce the predictors to a subset}
\usage{
geneSelection(
  data,
  covariate,
  t = 10,
  k = 10,
  n = 0.5,
  train.split = 0.8,
  sample.prob = c()
)
}
\arguments{
\item{data}{list of numeric expression matrices}

\item{covariate}{list of numeric vectors}

\item{t}{number of times to repeat the train/test split. Defaults to 10}

\item{k}{number of kfolds to execute in the GLMNET algorithm
(\link[glmnet]{cv.glmnet}). Defaults to 10}

\item{n}{numeric value. Minimum ratio of gene appearance to be selected as a
predictor}

\item{train.split}{numeric percentage of samples to take as train}

\item{sample.prob}{list of numeric vectors as weights when applying
\link[base]{sample}}
}
\value{
list of vectors containing the genes selected for each condition. If
only one data matrix is provided, it will return the reduced matrix, not a
list.
}
\description{
Estimates the most common subset of genes that the GLMNET algorithm considers
important to predict a covariate. The inputs can be either a list object
generated from previous SuCoNets functions or individual condition variables.
}
