% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryf.R
\name{get.r2}
\alias{get.r2}
\title{Getting the r^2 value for a prediction}
\usage{
get.r2(real.values, pred.values)
}
\arguments{
\item{real.values}{numeric vector of real values of the covariate}

\item{pred.values}{numeric vector of predicted values of the covariate}
}
\value{
The calculated r^2
}
\description{
Getting the r^2 value for a prediction
}
