% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get.r2_adj}
\alias{get.r2_adj}
\title{Getting the adjusted r^2 value for a prediction}
\usage{
get.r2_adj(r2, n, p)
}
\arguments{
\item{r2}{numeric, r2 value of the prediction}

\item{n}{number of samples}

\item{k}{number of predictors}
}
\value{
The calculated adjuted r^2
}
\description{
Getting the adjusted r^2 value for a prediction
}
