% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryf.R
\name{get.RMSE}
\alias{get.RMSE}
\title{Getting the RMSE for a model and a dataset}
\usage{
get.RMSE(real.values, pred.values)
}
\arguments{
\item{real.values}{numeric vector of real values of the covariate}

\item{pred.values}{numeric vector of predicted values of the covariate}
}
\value{
The calculated adjuted RMSE
}
\description{
Getting the RMSE for a model and a dataset
}
