% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSelection.R
\name{glmnetGenes}
\alias{glmnetGenes}
\title{Run GLMNET algorithm with the best seed}
\usage{
glmnetGenes(data, covariate, semilla)
}
\arguments{
\item{data}{expression matrix}

\item{covariate}{numeric vector}

\item{semilla}{number}
}
\value{
The generated glmnet object
}
\description{
Run GLMNET algorithm with the best seed
}
