% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSelection.R
\name{glmnetGenesSubset}
\alias{glmnetGenesSubset}
\title{Run GLMNET algorithm with the selected genes}
\usage{
glmnetGenesSubset(data, covariate, genes.subset)
}
\arguments{
\item{data}{expression matrix}

\item{covariate}{numeric vector}

\item{predictor.genes}{dataframe containing the gene selection and their frequency}
}
\value{
The generated glmnet object
}
\description{
Run GLMNET algorithm with the selected genes
}
