% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{groupMetrics}
\alias{groupMetrics}
\title{Summarize a metric across all bootstrap repetitions}
\usage{
groupMetrics(
  df,
  metrics = c("rmse.test", "Returned.predictors", "r2_adj.train")
)
}
\arguments{
\item{df}{numerical matrix (or list of numerical matrices) with predictors
as rows and samples as columns}

\item{covariate}{list of numeric vectors}

\item{seed}{fixed seed to generate split the dataset in train/test}
}
\value{
list by folds of data.train, data.test, covariate.train and covariate.test for
every condition given.
}
\description{
Summarize a metric across all bootstrap repetitions
}
