% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{histGeneFreq}
\alias{histGeneFreq}
\title{Histogram of gene frequencies}
\usage{
histGeneFreq(data, covariate, selectedGenes)
}
\arguments{
\item{data}{expression matrix}

\item{covariate}{numeric vector}

\item{selectedGenes}{dataframe with the genes selected as important by GLMNET algorithm}
}
\value{
Summary of gene frequency information
}
\description{
Histogram of gene frequencies
}
