% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryf.R
\name{indiceJaccard}
\alias{indiceJaccard}
\title{Calculate the Jaccard index}
\usage{
indiceJaccard(df, tam)
}
\arguments{
\item{df}{dataframe}

\item{tam}{size of the sets}
}
\value{
summary of Jaccard indexes
}
\description{
Given a dataframe df formed by sets of genes, this function calculates the Jaccard index
of each set with the rest and returns a summary with the information of all the calculated indexes.
Among the information returned by this function is the median of all the indexes.
The closer the median is to zero, the more independent the sets will be.
}
