% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{jaccardIndex}
\alias{jaccardIndex}
\title{Calculates the jaccard index of two sets}
\usage{
jaccardIndex(a, b)
}
\arguments{
\item{a}{set of elements of the first vector.}

\item{b}{set of elements of the second vector.}
}
\value{
jaccard index between the two sets.
}
\description{
Calculates the jaccard index of two sets
}
