% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plotAllSimilarity}
\alias{plotAllSimilarity}
\title{ToDo}
\usage{
plotAllSimilarity(
  genes.freq,
  heatplots,
  diag = "default",
  show.text = T,
  output.path = NA,
  draw.plot = T
)
}
\arguments{
\item{genes.freq}{data.frame obtained from the function \link[CovCoExpNets]{groupMetrics}}

\item{heatplots}{data.frame obtained from the function \link[CovCoExpNets]{groupMetrics}}

\item{diag}{(optinal) which parameter to register in the diagional. Only
other possible value is "genes". By default, the jaccard index.}

\item{show.text}{string, title of the plot}

\item{output.path}{string, y-axis title}

\item{draw.plot}{string, title of the legend}
}
\value{
ToDo
}
\description{
ToDo
}
