% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plotCoincidences}
\alias{plotCoincidences}
\title{Heatmap of the coincidences in genes selected for each condition}
\usage{
plotCoincidences(selected.genes)
}
\arguments{
\item{selected.genes}{list of dataframes with the genes selected as important
by GLMNET algorithm}
}
\value{
ggplot2 object
}
\description{
The inputs can be either a list object generated from previous SuCoNets
functions or individual condition variables.
}
