% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plotComparisonReturnedPredictors}
\alias{plotComparisonReturnedPredictors}
\title{Plots the returned predictors}
\usage{
plotComparisonReturnedPredictors(
  df.group.A,
  df.group.B,
  output.file = NA,
  title = "",
  ylab = "",
  legend.title = "Algorithm",
  first_arg = "GLMNET",
  second_arg = "CovCoExpNets"
)
}
\arguments{
\item{df.group.A}{data.frame obtained from the function \link[CovCoExpNets]{groupMetrics}}

\item{df.group.B}{data.frame obtained from the function \link[CovCoExpNets]{groupMetrics}}

\item{title}{string, title of the plot}

\item{ylab}{string, y-axis title}

\item{legend.title}{string, title of the legend}

\item{output_file}{path the output file to save the plot. If not provided, the plot will
not be stored}

\item{first.arg}{string, group name of df.group.A}

\item{second.arg}{string, group name of df.group.B}
}
\value{
ggplot object containing the returned predictors comparison of the two groups
}
\description{
For the two given groups (see \link[CovCoExpNets]{groupMetrics}),
it plots the returned predictors for the conditions and its 95\% confidence
interval.
}
