% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plotR2adj}
\alias{plotR2adj}
\title{Bar Plot of the RMSE values for each tissue (DEPRECATED)}
\usage{
plotR2adj(
  data,
  covariate,
  genes.subset,
  B = 20,
  sample.prob = c(),
  k = 10,
  plot.split = "test",
  do.boxplot = FALSE
)
}
\arguments{
\item{B}{number of bootstrap resamples (default 20)}

\item{tissues}{list of strings containing the tissue names}

\item{dataset_folder}{path to the dataset folder}

\item{models_folder}{path to the models folder}
}
\value{
plot
}
\description{
Bar Plot of the RMSE values for each tissue (DEPRECATED)
}
