% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubGenesDetection.R
\name{reduceGenes}
\alias{reduceGenes}
\title{Reduce the predictors to a subset}
\usage{
reduceGenes(genes.freq, mrfa = 0.9, force.mrfa = T, relative = T)
}
\arguments{
\item{genes.freq}{list of vectors containing the genes selected for each condition.
Shold be the output of \link[CovCoExpNets]{geneFrequency}}

\item{mrfa}{numeric, either the minimum frequency of appearance of a gene (0 <= mrda <= 1)
or the minimum number of times that a gene must appear in the
glmnet repetitions to be extracted (1 < mrfa < t)}

\item{force.mrfa}{boolean, whether to force the mrfa to be the given value. If set
to TRUE, the algorithm will keep decreasing the mrfa if no genes are being extracted}

\item{relative}{whether to use the relative mrfa. If TRUE, the parameter will be considered
as a percentage from 0 to 1. If FALSE, the parameter will be considered the plain number
of times the predictor must appear from 1 to t. By default, TRUE.}
}
\value{
list of vectors containing the genes selected for each condition. If
only one data matrix is provided, it will return the reduced matrix, not a
list.
}
\description{
Estimates the most common subset of genes that the GLMNET algorithm considers
important to predict a covariate. The input must be the output of
\link[CovCoExpNets]{geneFrequency}
}
