% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpreproc.R
\name{rRedundantPredictors}
\alias{rRedundantPredictors}
\title{Remove redundant predictors (DEPRECATED)}
\usage{
rRedundantPredictors(data)
}
\arguments{
\item{data}{numerical matrix with predictors as rows and samples as columns}
}
\value{
numerical matrix
}
\description{
Reduces the number of predictors by eliminating those with near-zero variance
and from pairs that have high correlation (>= 90\%) eliminate one of the two
predictors.
}
\examples{
nycfl = nycflights13::flights
# We won´t use those non numerical predictors
nycfl = na.omit(nycfl[,-c(1,10,12,13,14,19)])
ncol(nycfl)
nycfl2 <- t(rRedundantPredictors(t(nycfl)))
ncol(nycfl2)
}
