% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryf.R
\name{running.gprofiler}
\alias{running.gprofiler}
\title{Running the function gprofiler}
\usage{
running.gprofiler(selectedGenes, tam, data, df)
}
\arguments{
\item{selectedGenes}{dataframe with the genes selected as important by GLMNET algorithm}

\item{tam}{numeric vector containing the number of genes forming each cluster}

\item{data}{coexpression matrix}

\item{df}{dataframe with clusters}
}
\value{
The information obtained by running gprofiler
}
\description{
Running the function gprofiler
}
