% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpreproc.R
\name{scn}
\alias{scn}
\title{Scaling, centering and normalizing a matrix
Scales a non-negative numerical matrix by applying the logarithm function
and centers and normalizes its data. When applying the logarithm function
it adds one to avoid calculating the logarithm of zero.}
\usage{
scn(data)
}
\arguments{
\item{data}{non-negative numerical matrix with column names}
}
\value{
numerical matrix
}
\description{
Scaling, centering and normalizing a matrix
Scales a non-negative numerical matrix by applying the logarithm function
and centers and normalizes its data. When applying the logarithm function
it adds one to avoid calculating the logarithm of zero.
}
\examples{
m <- matrix(1:4, nr= 2, nc = 2)
colnames(m)<-c("1", "2")
scn(matrix(1:4, nr= 2, nc = 2))
}
