% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryf.R
\name{seleccionarFilas}
\alias{seleccionarFilas}
\title{Selects the indexes of the rows in x that match the rows in y}
\usage{
seleccionarFilas(x, y)
}
\arguments{
\item{x}{char vector}

\item{y}{char vector}
}
\value{
numeric vector
}
\description{
Selects the indexes of the rows in x that match the rows in y
}
