% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulateBivariate}
\alias{simulateBivariate}
\title{Generate a bivariate simulation}
\usage{
simulateBivariate(
  N,
  P,
  Q1,
  Q2,
  Qc,
  distribution = "uniform",
  params = list(min = -2, max = 2),
  seed = sample(1:99999999, 1),
  share.coeffs = F
)
}
\arguments{
\item{N}{number of samples to simulate.}

\item{P}{number of predictors to simulate.}

\item{Q1}{number of relevant predictors for the first covariate.}

\item{Q2}{number of relevant predictors for the second covariate.}

\item{Qc}{number of common relevant predictors between the two predictors. It must
be smaller than both Q1 and Q2.}

\item{distribution}{string indicating the distribution from which to take the
coefficients. Only uniform and gaussian distributions are possible. By default uniform}

\item{params}{list with the parameters from which to generate the coefficients. If
the distribution is uniform, it requires the minimum and maximum values. If the
distribution is gaussian, it requires the mean and standard deviation.}

\item{seed}{numeric seed to ensure reproducibility}

\item{share.coeffs}{boolean to indicate if the common predictors between the two
covariates should have the same coefficients. By default FALSE}
}
\value{
simulation object from where we can extract the data matrix and the two
response vectors.
}
\description{
Generates the simulation object. It includes the data matrix, the numeric
response vectors and additional information, like the coefficients of the
relevant predictors for each covariate. It is based on the package \link[simglm]{simglm}.
}
