% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulateUnivariate}
\alias{simulateUnivariate}
\title{Generate an univariate simulation}
\usage{
simulateUnivariate(
  N,
  P,
  Q,
  distribution = "uniform",
  params = list(min = -2, max = 2),
  seed = sample(1:99999999, 1)
)
}
\arguments{
\item{N}{number of samples to simulate.}

\item{P}{number of predictors to simulate.}

\item{Q}{number of relevant predictors.}

\item{distribution}{string indicating the distribution from which to take the
coefficients. Only uniform and gaussian distributions are possible. By default uniform.}

\item{params}{list with the parameters from which to generate the coefficients. If
the distribution is uniform, it requires the minimum and maximum values. If the
distribution is gaussian, it requires the mean and standard deviation.}

\item{seed}{numeric seed to ensure reproducibility.}
}
\value{
simulation object from where we can extract the data matrix and the
response vector.
}
\description{
Generates the simulation object. It includes the data matrix, the numeric
response vector and additional information, like the coefficients of the
relevant predictors. It is based on the package \link[simglm]{simglm}. It is
a wrapper of the \link[CovCoExpNets]{simulateBivariate} function.
}
