% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulateUnivariateLarge}
\alias{simulateUnivariateLarge}
\title{Generate an univariate simulation with large number of predictors}
\usage{
simulateUnivariateLarge(
  N,
  P,
  Q,
  distribution = "uniform",
  params = list(min = -2, max = 2),
  seed = sample(1:99999999, 1),
  max.predictors = 15000
)
}
\arguments{
\item{N}{number of samples to simulate.}

\item{P}{number of predictors to simulate.}

\item{Q}{number of relevant predictors.}

\item{distribution}{string indicating the distribution from which to take the
coefficients. Only uniform and gaussian distributions are possible. By default uniform.}

\item{params}{list with the parameters from which to generate the coefficients. If
the distribution is uniform, it requires the minimum and maximum values. If the
distribution is gaussian, it requires the mean and standard deviation.}

\item{seed}{numeric seed to ensure reproducibility.}

\item{max.predictors}{numeric maximum number of predictors per chunk of simulation.
By default, 15000.}
}
\value{
simulation object from where we can extract the data matrix and the
response vector.
}
\description{
Generates the simulation object when there are too many predictors for R to
handle properly. It splits the generation process in chunks of "max.predictors"
(by default, 15000). The main limitation is the formula size, which is limited
by default to a maximum value of arround 17000. Because of these limitations,
the relevant predictors will only be present in the first "max.predictors". That is,
if 1000 predictors are marked as relevant, they will be by default in the first 15000
predictors generated. Thus, the maximum number of relevant predictors is the argument
max.predictors. For more information about the simulation process, please refer to
\link[CovCoExpNets]{simulateUnivariate}. It is a wrapper of the
\link[CovCoExpNets]{simulateBivariateLarge} function.
}
