% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{splitByCondition}
\alias{splitByCondition}
\title{Splits data matrix by condition}
\usage{
splitByCondition(data, bool.matrix, names = NA)
}
\arguments{
\item{data}{numerical matrix with predictors as rows and samples as columns}

\item{bool.matrix}{boolean matrix with samples as rows and conditions as
columns. The data matrix will be splitted according to this boolean matrix}

\item{names}{names assigned to each different numerical matrix in the
returned list. If not provided, will be selected from the 'bool.matrix'}
}
\value{
named list of numerical matrices
}
\description{
Creates a lists of numeric matrices. Each matrix corresponds to a column in
the 'bool.matrix'. The names of the conditions can be provided or
extracted automatically from the 'bool.matrix' column names.
}
