% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{splitDataBootstrap}
\alias{splitDataBootstrap}
\title{Split data and covariate with bootstrapping}
\usage{
splitDataBootstrap(data, covariate, seed = sample(1:9999999, 1))
}
\arguments{
\item{data}{numerical matrix (or list of numerical matrices) with predictors
as rows and samples as columns}

\item{covariate}{list of numeric vectors}

\item{seed}{fixed seed to generate split the dataset in train/test}
}
\value{
named list with data.train, data.test, covariate.train and covariate.test
for every condition given.
}
\description{
Given a data matrix and a covariate vector (or a list of them), it
splits them into training set and testing set using a bootstrapping technique
}
