% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSelection.R
\name{stabilitySelection}
\alias{stabilitySelection}
\title{Study of the stability of the selection of genes with which we are going to work.}
\usage{
stabilitySelection(data, covariate, selectedGenes)
}
\arguments{
\item{data}{expression matrix}

\item{covariate}{numeric vector}

\item{selectedGenes}{dataframe with the genes selected as important by GLMNET algorithm}
}
\value{
Summary of gene frequency information
}
\description{
Study of the stability of the selection of genes with which we are going to work.
}
