% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{summaryVars}
\alias{summaryVars}
\title{Summarize a metric across all repetitions}
\usage{
summaryVars(
  df = NULL,
  variable,
  groups = NULL,
  conf.interval = 0.95,
  na.rm = T
)
}
\arguments{
\item{data}{numerical matrix (or list of numerical matrices) with predictors
as rows and samples as columns}

\item{covariate}{list of numeric vectors}

\item{seed}{fixed seed to generate split the dataset in train/test}
}
\value{
list by folds of data.train, data.test, covariate.train and covariate.test for
every condition given.
}
\description{
Summarize a metric across all repetitions
}
