% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{toOneHotEncoding}
\alias{toOneHotEncoding}
\title{Converts a list of predictors into a One Hot Encoding list}
\usage{
toOneHotEncoding(all.names, selected.names)
}
\arguments{
\item{all.names}{the total list of predictors.}

\item{selected.names}{the predictors that are going to be encoded.}
}
\value{
named vector of 1s and 0s
}
\description{
Converts a list of predictors into a One Hot Encoding list
}
