% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpreproc.R
\name{applyNearZeroVar}
\alias{applyNearZeroVar}
\title{Remove near zero variance predictors}
\usage{
applyNearZeroVar(data)
}
\arguments{
\item{data}{numerical matrix with predictors as rows and samples as columns}
}
\value{
reduced numerical matrix
}
\description{
Identifies and eliminates predictors that have very few unique values and the
ratio of the frecuency of the most common value to the frequency of the
second most common value is large. More info at \link[caret]{nearZeroVar}.
}
