% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{applyProteinCodingRequirement}
\alias{applyProteinCodingRequirement}
\title{Remove non protein-coding genes}
\usage{
applyProteinCodingRequirement(data)
}
\arguments{
\item{data}{numerical matrix with predictors as rows and samples as columns}
}
\value{
reduced numerical matrix
}
\description{
Identifies and eliminates genes which are not protein coding.
}
