% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryf.R
\name{calculateClusters}
\alias{calculateClusters}
\title{Calculates clusters by varying the size}
\usage{
calculateClusters(x, y, covariate)
}
\arguments{
\item{x}{coexpression matrix}

\item{y}{factor}

\item{covariate}{numeric vector}
}
\value{
dataframe with clusters
}
\description{
Calculates clusters by varying the size
}
