% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryf.R
\name{calculateClusters}
\alias{calculateClusters}
\title{Calculates clusters by varying the size}
\usage{
calculateClusters(x, y, cjt, covariate, seed)
}
\arguments{
\item{x}{coexpression matrix}

\item{y}{factor}

\item{cjt}{name of the genes selected as important by GLMNET algorithm}

\item{covariate}{numeric vector}

\item{seed}{number}
}
\value{
dataframe with clusters
}
\description{
Calculates clusters by varying the size
}
