% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSelection.R
\name{coexpressionNetworkFixed}
\alias{coexpressionNetworkFixed}
\title{Co-expression network with fixed cluster sizes}
\usage{
coexpressionNetworkFixed(data, genes, tam)
}
\arguments{
\item{data}{expression matrix}

\item{genes}{dataframe with the genes selected by the glmnet algorithm and their coefficients}

\item{tam}{number: indicates the number of genes in each cluster}
}
\value{
A dataframe where for each gene selected by glmnet appears the tam genes most correlated with it and that correlation
}
\description{
Calculate a coexpression network using the genes selected by the glmnet algorithm
as the best predictors of a covariate.
}
