% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSelection.R
\name{coexpressionNetworkVariable}
\alias{coexpressionNetworkVariable}
\title{Co-expression network with variable cluster sizes}
\usage{
coexpressionNetworkVariable(data, selectedGenes, covariate, seed)
}
\arguments{
\item{data}{expression matrix}

\item{selectedGenes}{dataframe with the genes selected by the glmnet algorithm and their coefficients}

\item{covariate}{numeric vector}

\item{seed}{number}
}
\value{
A dataframe where for each gene selected by glmnet appears the tam genes most correlated with it and a vector with cluster sizes
}
\description{
Co-expression network with variable cluster sizes
}
