% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{comparisonActualPredictedCovariate}
\alias{comparisonActualPredictedCovariate}
\title{Comparison between real values of the covariate and predicted values}
\usage{
comparisonActualPredictedCovariate(
  data,
  covariate,
  m,
  d,
  cvfit,
  genes.subset,
  sample.prob = c()
)
}
\arguments{
\item{data}{list of expression matrices}

\item{covariate}{list of numeric vectors}

\item{m}{list of means of the original covariates}

\item{d}{list of standard deviations of the original covariates}

\item{cvfit}{list of objects generated with GLMNET algorithm}

\item{genes.subset}{list of selected predictors for each condition}

\item{sample.prob}{list of numeric vectors as weights when applying
\link[base]{sample}}
}
\value{
plot
}
\description{
The inputs can be either a list object generated from previous SuCoNets
functions or individual condition variables.
}
\examples{
comparisonActualPredictedCovariate(data, covariate, m, d, cvfit, genes.subset)
comparisonActualPredictedCovariate(data[[1]], covariate[[1]], m[[1]], d[[1]], cvfit[[1]], genes.subset[[1]])
}
