% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{dataPreprocess}
\alias{dataPreprocess}
\title{Preprocess the input data matrices}
\usage{
dataPreprocess(
  data,
  doLog2 = T,
  doAT = T,
  doNZV = T,
  doGTExCorr = T,
  doProtCod = T,
  doZSco = T,
  doCorr = F,
  threshold = 0.1,
  perc = 0.8,
  freqCut = 20,
  uniqueCut = 5,
  sigDigits = 1,
  by.columns = F,
  cutoff = 0.9
)
}
\arguments{
\item{data}{numerical matrix (or list of numerical matrices) with predictors
as rows and samples as columns}

\item{doLog2}{boolean to apply \link[CovCoExpNets]{applyLog2}}

\item{doAT}{boolean to apply \link[CovCoExpNets]{applyActiveThreshold}}

\item{doNZV}{boolean to apply \link[CovCoExpNets]{applyNearZeroVar}}

\item{doGTExCorr}{boolean to apply \link[CovCoExpNets]{correctGTExNames}}

\item{doProtCod}{boolean to apply \link[CovCoExpNets]{applyProteinCodingRequirement}}

\item{doZSco}{boolean to apply \link[CovCoExpNets]{applyZScore}}

\item{doCorr}{boolean to apply \link[CovCoExpNets]{applyCorrSkew}}

\item{threshold}{numeric, required for \link[CovCoExpNets]{applyActiveThreshold}. minimum TPM value to consider a predictor 'active'.
Defaults to 0.1}

\item{perc}{numeric, required for \link[CovCoExpNets]{applyActiveThreshold}. minimum percentage of activated predictors in all
samples. Defaults to 0.8}

\item{freqCut}{numerical, required for \link[CovCoExpNets]{applyNearZeroVar}. The cutoff for the ratio of the most common value to the second most common value}

\item{uniqueCut}{numerical, required for \link[CovCoExpNets]{applyNearZeroVar}. The cutoff for the percentage of distinct values out of the number of total samples}

\item{sigDigits}{numerical, required for \link[CovCoExpNets]{applyNearZeroVar}. Number of significant digits to round}

\item{by.columns}{boolean. Whether to normalize by columns of rows, by default FALSE}

\item{cutoff}{numeric: value for the pair-wise absolute correlation cutoff (\link[caret]{findCorrelation}). Defaults to 0.9}
}
\value{
named list of reduced numerical matrices. If only one data matrix is
provided, it will return the reduced matrix, not a list.
}
\description{
Applies all the preprocessing required for the CovCoExpNets pipeline to work.
Each individual process can be deactivadted (but not recommended).
}
\details{
If input data is from the GTEx foundation, it is requiered to execute the naming
correction to process the protein coding genes.
}
