% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{distributionIndividualsCovariate}
\alias{distributionIndividualsCovariate}
\title{Distribution of individuals according to the covariate}
\usage{
distributionIndividualsCovariate(data, covariate, selectedGenes, m, d)
}
\arguments{
\item{data}{expression matrix}

\item{covariate}{numeric vector}

\item{selectedGenes}{dataframe with the genes selected as important by GLMNET algorithm}

\item{m}{mean of the original covariate}

\item{d}{standard deviation of the original covariate}
}
\value{
plot
}
\description{
Distribution of individuals according to the covariate
}
