% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubGenesDetection.R
\name{extractModelGenes}
\alias{extractModelGenes}
\title{Detect main genes}
\usage{
extractModelGenes(cvfit, genes.freq = NA)
}
\arguments{
\item{cvfit}{list of GMLNET object}
}
\value{
list of dataframe containing the variables selected by the glmnet
algorithm and the coefficients associated with these variables. If only one
object was supplied, the output will be one object and not a list.
}
\description{
This function calculates the variables that best predict a covariate using
glmnet algorithm. The inputs can be either a list object generated from
previous SuCoNets functions or individual condition variables.
}
