% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSelection.R
\name{geneSelection}
\alias{geneSelection}
\title{Estimates the most common subset of genes to predict a covariate}
\usage{
geneSelection(data, covariate, n = 0.5)
}
\arguments{
\item{data}{expression matrix}

\item{covariate}{numeric vector}

\item{n}{numeric value. Minimum ratio of appearances to be selected}
}
\value{
Dataframe containing the genes selected and their frequency
}
\description{
Estimates the most common subset of genes to predict a covariate
}
