% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSelection.R
\name{geneSelectionBootstrap}
\alias{geneSelectionBootstrap}
\title{Estimates the most common subset of genes to predict a covariate using Bootstrap}
\usage{
geneSelectionBootstrap(data, covariate, B = 20, n = 0.5)
}
\arguments{
\item{data}{expression matrix}

\item{covariate}{numeric vector}

\item{B}{number of bootstrap resamples (default 20)}

\item{n}{numeric value. Minimum ratio of appearances to be selected}
}
\value{
List containing the genes selected and their frequency
}
\description{
Estimates the most common subset of genes to predict a covariate using Bootstrap
}
