% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryf.R
\name{get.r2_adj}
\alias{get.r2_adj}
\title{Getting the adjusted r^2 value for a prediction}
\usage{
get.r2_adj(real.values, pred.values, k)
}
\arguments{
\item{real.values}{numeric vector of real values of the covariate}

\item{pred.values}{numeric vector of predicted values of the covariate}

\item{k}{number of predictors (length of selected genes)}

\item{n}{number of samples}
}
\value{
The calculated adjuted r^2
}
\description{
Getting the adjusted r^2 value for a prediction
}
