% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSelection.R
\name{glmnetGenesSubset}
\alias{glmnetGenesSubset}
\title{Run GLMNET algorithm with the selected genes}
\usage{
glmnetGenesSubset(
  data,
  covariate,
  genes.subset,
  k = 10,
  train.split = 0.8,
  sample.prob = c()
)
}
\arguments{
\item{data}{list of numeric expression matrices}

\item{covariate}{list of numeric vectors}

\item{genes.subset}{list of selected predictors for each condition}

\item{k}{number of kfolds to execute in the GLMNET algorithm
(\link[glmnet]{cv.glmnet}). Defaults to 10}

\item{train.split}{numeric percentage of samples to take as train}

\item{sample.prob}{list of numeric vectors as weights when applying
\link[base]{sample}}
}
\value{
list of the generated glmnet object. If only one data matrix is
provided, it will return only one object, not a list.
}
\description{
The inputs can be either a list object generated from previous SuCoNets
functions or individual condition variables.
}
