% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubGenesDetection.R
\name{glmnetGenesSubset}
\alias{glmnetGenesSubset}
\title{Run GLMNET algorithm with the selected genes}
\usage{
glmnetGenesSubset(
  data,
  covariate,
  genes.subset,
  k.folds = 10,
  train.split = 1,
  sample.prob = c(),
  seed = sample(1:999999, 1),
  glmnet.family = "gaussian",
  evaluate.model = T,
  m = NA,
  d = NA
)
}
\arguments{
\item{data}{list of numeric expression matrices}

\item{covariate}{list of numeric vectors}

\item{genes.subset}{list of selected predictors for each condition}

\item{k.folds}{number of kfolds to execute in the GLMNET algorithm
(\link[glmnet]{cv.glmnet}). Defaults to 10}

\item{train.split}{numeric percentage of samples to take as train}

\item{sample.prob}{list of numeric vectors as weights when applying
\link[base]{sample}}

\item{seed}{fixed seed to generate split the dataset in train/test}

\item{evaluate.model}{whether to evaluate the generated models
using train/test splits with the train.split argument.}

\item{m}{list of means for the covariate}

\item{d}{list of standard deviations for the covariate}
}
\value{
list of the generated glmnet object. If only one data matrix is
provided, it will return only one object, not a list. If the evaluate
argument is set to TRUE, it will also return the results of
evaluating the model with the test dataset provided.
}
\description{
The inputs can be either a list object generated from previous CovCoExpNets
functions or individual condition variables. Aditionally, a list of
predictors must be provided to prune the training datsets.
}
