% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{histGeneFreq}
\alias{histGeneFreq}
\title{Histogram of gene frequencies}
\usage{
histGeneFreq(gs.our.genes)
}
\arguments{
\item{gs.our.genes}{data.frame}
}
\value{
plot
}
\description{
Histogram of gene frequencies
}
