% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plotCoincidences}
\alias{plotCoincidences}
\title{Heatmap of the coincidences in genes selected for each tissue}
\usage{
plotCoincidences(tissues, dataset_folder, models_folder)
}
\arguments{
\item{tissues}{list of strings containing the tissue names}

\item{dataset_folder}{path to the dataset folder}

\item{models_folder}{path to the models folder}
}
\value{
plot
}
\description{
Heatmap of the coincidences in genes selected for each tissue
}
