% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plotRMSE}
\alias{plotRMSE}
\title{Bar Plot of the RMSE values for each condition}
\usage{
plotRMSE(
  data,
  covariate,
  genes.subset,
  m = 0,
  d = 1,
  B = 20,
  sample.prob = c(),
  k = 10,
  do.boxplot = FALSE
)
}
\arguments{
\item{data}{list of expression matrices}

\item{covariate}{list of numeric vectors}

\item{genes.subset}{list of selected predictors for each condition}

\item{B}{number of bootstrap resamples. Defaults to 20}

\item{sample.prob}{list of numeric vectors as weights when applying
\link[base]{sample}}

\item{k}{number of kfolds to execute in the GLMNET algorithm
(\link[glmnet]{cv.glmnet}). Defaults to 10}

\item{do.boxplot}{boolean to do a boxplot or a barplot. Defaults to False}
}
\value{
ggplot2 object
}
\description{
The inputs can be either a list object generated from previous SuCoNets
functions or individual condition variables.
}
