% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryf.R
\name{running.gprofiler}
\alias{running.gprofiler}
\title{Running the function gprofiler}
\usage{
running.gprofiler(selectedGenes, data, df, tam)
}
\arguments{
\item{selectedGenes}{dataframe or list of dataframes, with the genes selected
as important by GLMNET algorithm.}

\item{data}{coexpression matrix. Also accepts a list of coexpression matrices
for different conditions.}

\item{df}{dataframe with clusters. Also accepts a list of dataframes for
different conditions.}

\item{tam}{number: indicates the number of genes in each cluster. Only
required for fixed networks. Must be the same as the one provided to
coexpressionNetworkFixed.}
}
\value{
The information obtained by running gprofiler
}
\description{
Running the function gprofiler
}
